Finished coordinate system for Quadratics

+++ Math.Solver.Core/Model/Navigation/Point.cs
@@ -2,13 +2,13 @@
 {
     public class Point
     {
-        public Point(int x, int y)
+        public Point(decimal x, decimal y)
         {
             X = x;
             Y = y;
 
         }
-        public int  X { get; set; }
-        public int Y { get; set; }
+        public decimal  X { get; set; }
+        public decimal Y { get; set; }
     }
 }
